--- src/components/Pages/FullWidthBackground.css
@@ -1,11 +1,0 @@
-.full-width-bg {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background: #1d2026;
-  margin: 0;
-  padding: 0;
-  z-index: -1;
-}
