+++ src/components/Pages/Profile.module.scss
@@ -1,19 +1,26 @@
 .profileContainer {
-  background: rgba(40,44,52,0.85);
+  background: rgba(30, 34, 40, 0.85);
   border-radius: 24px;
   box-shadow: 0 8px 32px rgba(0,0,0,0.18);
-  color: #61dafb;
+  color: #2da7f8;
   backdrop-filter: blur(8px);
   border: 1px solid rgba(97,218,251,0.12);
   padding: 2.7rem 2.2rem;
-  max-width: 440px;
+  max-width: 640px;
   margin: 2.7rem auto;
   font-family: 'Segoe UI', 'Roboto', Arial, sans-serif;
-  transition: box-shadow 0.5s, background 0.5s;
-  min-height: 100vh;
+  transition: box-shadow 0.5s, background 0.5s, color 0.5s;
+  animation: fadeInProfile 1.2s cubic-bezier(0.4, 0, 0.2, 1);
   overflow-y: auto;
+  position: relative;
+  display: flex;
+  flex-direction: column;
   align-items: stretch;
   justify-content: flex-start;
+}
+@keyframes fadeInProfile {
+  from { opacity: 0.6; transform: translateY(16px); }
+  to { opacity: 1; transform: translateY(0); }
 }
 .profileTitle {
   color: #61dafb;
@@ -24,7 +31,12 @@
   text-shadow: 0 2px 8px rgba(97,218,251,0.15);
 }
 .profileCard {
-  transition: transform 0.4s cubic-bezier(0.4, 0, 0.2, 1), box-shadow 0.4s;
+  transition: transform 0.4s cubic-bezier(0.4, 0, 0.2, 1), box-shadow 0.4s, color 0.4s;
+  animation: cardPop 1.2s cubic-bezier(0.4, 0, 0.2, 1);
+}
+@keyframes cardPop {
+  from { opacity: 0; transform: scale(0.95); }
+  to { opacity: 1; transform: scale(1); }
 }
 .profileCard:hover {
   transform: scale(1.03) translateY(-4px);
@@ -40,10 +52,10 @@
   padding: 0.8rem 1.2rem;
   box-shadow: 0 1px 4px rgba(0,0,0,0.08);
   transition: background 0.3s, color 0.3s, box-shadow 0.3s;
-  color: #61dafb;
+  color: #2da7f8;
 }
 .profileField:hover {
-  background: #61dafb;
+  background: #2da7f8;
   color: #181c22;
   box-shadow: 0 4px 16px #61dafb44;
   transform: scale(1.04);
@@ -51,6 +63,6 @@
 .profileLabel {
   font-weight: 700;
   margin-right: 0.8rem;
-  color: #61dafb;
+  color: #2da7f8;
   text-shadow: 0 2px 8px rgba(40,44,52,0.85);
 }
