+++ src/components/ProfileComponent/Footer.module.scss
@@ -1,11 +1,19 @@
 .footer {
-  padding: 1.2rem 2rem;
-  background: #20232a;
-  color: #fff;
+  padding: 2rem 12rem;
+  background: rgba(40,44,52,0.85);
+  color: #2da7f8;
   text-align: center;
   font-size: 1rem;
-  border-radius: 16px 16px 0 0;
-  box-shadow: 0 -2px 12px rgba(0,0,0,0.08);
+  border-radius: 0 0 24px 24px;
+  box-shadow: 0 -2px 16px rgba(0,0,0,0.10);
+  max-width: 540px;
+  margin: 2rem auto 0 auto;
   font-family: 'Segoe UI', 'Roboto', Arial, sans-serif;
-  transition: box-shadow 0.4s, background 0.4s;
+  transition: box-shadow 0.4s, background 0.4s, color 0.4s;
+  animation: fadeInFooter 1.2s cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+@keyframes fadeInFooter {
+  from { opacity: 0; transform: translateY(40px); }
+  to { opacity: 1; transform: translateY(0); }
 }
