+++ src/components/ProfileComponent/Header.jsx
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 
 export default function Header() {
   return (
-    <header className="header">
+    <header className={styles.header}>
       <nav className={styles.nav}>
         <Link to="/">Home</Link>
         <Link to="/profile">Profile</Link>
