+++ src/components/ProfileComponent/Hero.module.scss
@@ -1,40 +1,49 @@
+/* Explicit font size for profile card details */
+.profileCard {
+  font-size: 1.5rem;
+  transition: transform 0.4s cubic-bezier(0.4, 0, 0.2, 1), box-shadow 0.4s, color 0.4s;
+  animation: cardPopHero 1.2s cubic-bezier(0.4, 0, 0.2, 1);
+}
+@keyframes cardPopHero {
+  from { opacity: 0; transform: scale(0.95); }
+  to { opacity: 1; transform: scale(1); }
+}
 .hero {
-  padding: 3rem 2rem;
+  padding: 2.5rem 6rem 6rem 6rem;
   text-align: center;
-  background: #20232a;
-  border-radius: 16px;
-  box-shadow: 0 4px 24px rgba(0,0,0,0.08);
-  margin: 2rem auto;
-  max-width: 600px;
+  background: rgba(40,44,52,0.85);
+  border-radius: 24px;
+  box-shadow: 0 4px 24px rgba(0,0,0,0.10);
+  margin: auto auto;
+  max-width: 540px;
   font-family: 'Segoe UI', 'Roboto', Arial, sans-serif;
   color: #fff;
-  transition: box-shadow 0.4s, background 0.4s, color 0.3s;
+  transition: box-shadow 0.4s, background 0.4s, color 0.4s;
+  animation: fadeInHero 1.2s cubic-bezier(0.4, 0, 0.2, 1);
 }
-
-.hero:hover {
-  box-shadow: 0 12px 40px #61dafb33;
-  color: #21a1f1;
-  transform: scale(1.03);
+@keyframes fadeInHero {
+  from { opacity: 0; transform: translateY(40px); }
+  to { opacity: 1; transform: translateY(0); }
 }
-
 .hero h2 {
-  font-size: 2rem;
+  font-size: 3rem;
   color: #61dafb;
   font-weight: 700;
-  margin-bottom: 1rem;
+  margin-top: 0;
+  margin-bottom: 3rem;
 }
 .hero p {
-  font-size: 1.15rem;
+  font-size: 2rem;
   color: #fff;
-  margin-bottom: 1.5rem;
+  margin-bottom: 1rem;
 }
 .button {
   padding: 0.7rem 1.5rem;
   margin-top: 1rem;
   border: none;
   border-radius: 8px;
-  background: #61dafb;
-  color: #282c34;
+  background: #2da7f8;
+  color: #464c57;
   font-size: 1.1rem;
   font-weight: 600;
   cursor: pointer;
@@ -46,4 +55,7 @@
   color: #fff;
   box-shadow: 0 4px 16px #61dafb44;
   transform: translateY(-2px) scale(1.05);
+}
+.profileLabel {
+  color: #21a1f1; // Replace #yourColor with your desired color code
 }
