+++ src/components/ProfileComponent/Title.module.scss
@@ -16,10 +16,15 @@
   max-width: 600px;
   margin-left: auto;
   margin-right: auto;
-  transition: box-shadow 0.4s, background 0.4s, color 0.3s;
+  transition: color 0.4s, text-shadow 0.4s;
+  animation: fadeInTitle 1.2s cubic-bezier(0.4, 0, 0.2, 1);
 }
 .title:hover {
   box-shadow: 0 12px 40px #61dafb33;
   color: #21a1f1;
   transform: scale(1.03);
+}
+@keyframes fadeInTitle {
+  from { opacity: 0; transform: translateY(40px); }
+  to { opacity: 1; transform: translateY(0); }
 }
